Demo_AI/ai_code_reviewer/requirements.txt
openai >= 1.0.0
python-dotenv >= 0.19.0